Move resource usage merge types under line cap
